AUT-796 - Call the new PrincipalValidationHelper from the AM lambdas
- Remove the method from the RequestBodyHelper as we now handle this in the PrincipalValidationHelper
- Call the principleIsInvalid from the AM lambdas and if the principal is not valid then throw an exception
- Add extra unit tests so we cover both scenarios of when the principalId contains the publicSubjectId and when it contains the internal subject pairwise id.

+++ account-management-api/src/main/java/uk/gov/di/accountmanagement/lambda/RemoveAccountHandler.java
@@ -4,20 +4,20 @@ import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
-import com.nimbusds.oauth2.sdk.id.Subject;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import uk.gov.di.accountmanagement.domain.AccountManagementAuditableEvent;
 import uk.gov.di.accountmanagement.entity.NotificationType;
 import uk.gov.di.accountmanagement.entity.NotifyRequest;
 import uk.gov.di.accountmanagement.entity.RemoveAccountRequest;
+import uk.gov.di.accountmanagement.exceptions.InvalidPrincipalException;
+import uk.gov.di.accountmanagement.helpers.PrincipalValidationHelper;
 import uk.gov.di.accountmanagement.services.AwsSqsClient;
 import uk.gov.di.authentication.shared.entity.ErrorResponse;
 import uk.gov.di.authentication.shared.exceptions.UserNotFoundException;
 import uk.gov.di.authentication.shared.helpers.IpAddressHelper;
 import uk.gov.di.authentication.shared.helpers.LocaleHelper.SupportedLanguage;
 import uk.gov.di.authentication.shared.helpers.PersistentIdHelper;
-import uk.gov.di.authentication.shared.helpers.RequestBodyHelper;
 import uk.gov.di.authentication.shared.helpers.RequestHeaderHelper;
 import uk.gov.di.authentication.shared.serialization.Json;
 import uk.gov.di.authentication.shared.serialization.Json.JsonException;
@@ -107,9 +107,14 @@ public class RemoveAccountHandler
                                                     "User not found with given email"));
 
             Map<String, Object> authorizerParams = input.getRequestContext().getAuthorizer();
-            RequestBodyHelper.validatePrincipal(
-                    new Subject(userProfile.getPublicSubjectID()), authorizerParams);
 
+            if (PrincipalValidationHelper.principleIsInvalid(
+                    userProfile,
+                    configurationService.getInternalSectorUri(),
+                    authenticationService,
+                    authorizerParams)) {
+                throw new InvalidPrincipalException("Invalid Principal in request");
+            }
             authenticationService.removeAccount(email);
             LOG.info("User account removed. Adding message to SQS queue");
